Invert isAdmin before handing it to DeleteUserForm
Because the form is stupid, it's "impossible" to let it handle isActive as an inverted toggle.
Instead we now invert the value, then hand it to the form. When it comes back we handle it the same
way as before, but in its inverted state. I.e. true means that `isActive` is `false`.

See DIH-399